Added server logger

+++ src/events/messageCreate.ts
@@ -1,10 +1,15 @@
 import { Message } from "discord.js";
+import logger from "../logger/logger";
+import { IBotClient } from "../types";
 
 const forbiddenPhrases: string[] = ["discord.gg"];
 
 module.exports = {
     name: "messageCreate",
-    execute(message: Message) {
+    execute(message: Message, client: IBotClient, logger: logger) {
+
+        if (message.author.bot)
+            return;
         const foundPhrase = forbiddenPhrases.find((phrase) => message.content.includes(phrase));
         if(foundPhrase) return message.delete();
 
@@ -14,5 +19,7 @@ module.exports = {
         ) {
             message.delete();
         }
+
+        logger.messageCreateEvent(message, client);
     },
 };
